some more feature

+++ src/SiteBundle/Controller/OrderController.php
@@ -9,7 +9,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Component\HttpFoundation\Response;
-
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 class OrderController extends Controller
 {
@@ -20,6 +20,11 @@ class OrderController extends Controller
 
 
         $user = $this->getUser();
+        if (!is_object($user)) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
+
+
 
         $em    = $this->get('doctrine.orm.entity_manager');
         $qb = $em->createQueryBuilder();
@@ -37,9 +42,7 @@ class OrderController extends Controller
             10
         );
 
-        $translator=$this->get('translator');
-        $seoPage = $this->get('sonata.seo.page');
-        $seoPage->setTitle($translator->trans('sitebundle.htmltitle.order'));
+        $this->get('site.setSeo')->setTitle('sitebundle.yourorders');
 
 
         return $this->render('SiteBundle:Order:index.html.twig',array(
@@ -48,11 +51,15 @@ class OrderController extends Controller
         ));
     }
     /**
-     * @ParamConverter("get", class="LilWorksStoreBundle:Order", options={"id" = "id"})
+     * @ParamConverter("order", options={"mapping": {"order_id"   : "id"}})
      */
     public function pdfAction(Request $request,Order $order)
     {
 
+        $user = $this->getUser();
+        if (!is_object($user)  || $order->getCustomer() != $user->getCustomer()) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
         $em    = $this->getDoctrine()->getEntityManager();
 
         $param = array();
@@ -63,6 +70,10 @@ class OrderController extends Controller
             'order'  => $order,
             'base_dir' => $this->get('kernel')->getRootDir() . '/../web' . $request->getBasePath(),
             'param'=>$param
+        ));
+
+        $this->get('site.flash')->setMessages(array(
+            array('status'=>'success','message'=>'sitebundle.flash.pdf.created')
         ));
 
         return $this->render('SiteBundle:Order:pdf.html.twig',array(
@@ -94,49 +105,87 @@ class OrderController extends Controller
         );
     }
     /**
-     * @ParamConverter("get", class="LilWorksStoreBundle:Order", options={"id" = "id"})
+     * @ParamConverter("order", options={"mapping": {"order_id"   : "id"}})
      */
     public function editAction(Request $request,Order $order){
 
-        $translator=$this->get('translator');
-        $seoPage = $this->get('sonata.seo.page');
-        $seoPage->setTitle($translator->trans('sitebundle.htmltitle.order %reference%') , array('%reference%'=>$order->getReference() ));
 
+        $user = $this->getUser();
+        if (!is_object($user)  || $order->getCustomer() != $user->getCustomer()) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
+        $form = $this->createForm('SiteBundle\Form\OrderType', $order);
+
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $this->getDoctrine()->getManager()->persist($order);
+            $this->getDoctrine()->getManager()->flush();
+            $this->get('site.flash')->setMessages(array(
+                array('status'=>'success','message'=>'sitebundle.flash.order.updated')
+            ));
+        }
+        $this->get('site.setSeo')->setTitle('sitebundle.yourorder.edit %reference%',array('%reference%'=>$order->getReference()));
 
         return $this->render('SiteBundle:Order:edit.html.twig',array(
+            'order'=>$order,
+            'form'=>$form->createView()
+        ));
+    }
+    /**
+     * @ParamConverter("order", options={"mapping": {"order_id"   : "id"}})
+     */
+    public function showAction(Request $request,Order $order){
+
+        $user = $this->getUser();
+        if (!is_object($user)  || $order->getCustomer() != $user->getCustomer()) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
+
+        $this->get('site.setSeo')->setTitle('sitebundle.yourorder.show %reference%',array('%reference%'=>$order->getReference()));
+
+        return $this->render('SiteBundle:Order:show.html.twig',array(
             'order'=>$order
         ));
     }
     /**
-     * @ParamConverter("get", class="LilWorksStoreBundle:Order", options={"id" = "id"})
+     * @ParamConverter("order", options={"mapping": {"order_id"   : "id"}})
      */
     public function payAction(Request $request,Order $order){
 
+        $user = $this->getUser();
+        if (!is_object($user)  || $order->getCustomer() != $user->getCustomer()) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
         $paymentService = $this->get('site.payment');
         $paymentMethodsForm = array();
 
 
         $paymentMethods = $this->getDoctrine()->getRepository("LilWorksStoreBundle:PaymentMethod")->findBy(array('isPublished'=>1));
-        foreach($paymentMethods as $paymentMethod){
+        /*foreach($paymentMethods as $paymentMethod){
             $paymentService->setPaymentMethod($paymentMethod,$order);
             array_push($paymentMethodsForm,$paymentService->getForm());
-        }
+        }*/
 
-        $translator=$this->get('translator');
-        $seoPage = $this->get('sonata.seo.page');
-        $seoPage->setTitle($translator->trans('sitebundle.htmltitle.order.pay %reference%') , array('%reference%'=>$order->getReference() ));
-
+        $this->get('site.setSeo')->setTitle('sitebundle.yourorder.pay %reference%',array('%reference%'=>$order->getReference()));
 
         return $this->render('SiteBundle:Order:pay.html.twig',array(
             'order'=>$order,
             'paymentMethods'=>$paymentMethods,
-            'paymentMethodsForm'=>$paymentMethodsForm
+            'paymentService'=>$paymentService,
+            #'paymentMethodsForm'=>$paymentMethodsForm
         ));
     }
-
+    /**
+     * @ParamConverter("order", options={"mapping": {"order_id"   : "id"}})
+     */
     public function payedAction(Request $request,Order $order){
 
-        var_dump($order->getId());
+
+        $user = $this->getUser();
+        if (!is_object($user)  || $order->getCustomer() != $user->getCustomer()) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
         $em = $this->getDoctrine()->getEntityManager();
         $orderStep =  $em->getRepository("LilWorksStoreBundle:OrderStep")->findOneByTag("PAYED");
 
@@ -164,16 +213,33 @@ class OrderController extends Controller
         ));
     }
 
-
     /**
-     * @ParamConverter("get", class="LilWorksStoreBundle:Order", options={"id" = "id"})
+     * @ParamConverter("order", options={"mapping": {"order_id"   : "id"}})
      */
-    public function removeAction(Request $request,Order $order){
+    public function deleteAction(Request $request,Order $order){
 
-        $em = $this->getDoctrine()->getEntityManager();
-        $em->remove($order);
-        $em->flush();
-        return $this->redirectToRoute('site_order');
+        $user = $this->getUser();
+        if (!is_object($user)  || $order->getCustomer() != $user->getCustomer()) {
+            throw new AccessDeniedException('This user does not have access to this section.');
+        }
+
+        if($this->get('lilworks.store.orderManager')->isRemovable($order)){
+            $em = $this->getDoctrine()->getEntityManager();
+            $em->remove($order);
+            $em->flush();
+
+            $this->get('site.flash')->setMessages(array(
+                array('status'=>'success','message'=>'sitebundle.flash.order.deleted')
+            ));
+        }else{
+            $this->get('site.flash')->setMessages(array(
+                array('status'=>'error','message'=>'sitebundle.flash.order.notdeleted')
+            ));
+        }
+
+
+        return $this->redirectToRoute('site_orders');
     }
+
 
 }
